pyproject.toml
[tool.poetry]
name = "openai-sandbox-py"
version = "0.1.0"
description = ""
authors = ["Juan Guerrero <JCystems@GMail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
json5 = "^0.9.14"
langchain = {extras = ["all"], version = "0.1.0"}
matplotlib = "^3.7.1"
num2words = "^0.5.12"
openai = "^1.10.0"
pandas = "^2.0.2"
plotly = "^5.14.1"
pydantic = {extras = ["dotenv"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
redis = "^4.5.5"
scikit-learn = "^1.2.2"
semantic-kernel = "0.4.7.dev"
tiktoken = "^0.4.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^24.1.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pyclean = "^2.7.6"
pyhamcrest = "^2.1.0"
pylint = "^2.17.4"
pyright = "^1.1.348"
pytest = "^8.3.5"
pytest-dotenv = "^0.5.2"
ruff = "^0.1.14"

[tool.autoflake]
check = true

[tool.black]
line-length = 150
target-version = ["py311"]

[tool.pyright]
venvPath = "."
venv = ".venv"
defineConstant = { DEBUG = true }
typeCheckingMode = "strict"
include = ["src"]
strict = ["src"]
stubPath = "typings"
exclude = [
    ".venv",
    "notebooks",
    "typings",
    "**/__pycache__",
    "**/node_modules",
    "tests/**/test*.py",
    "XXX",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
